Routine activities delete

+++ api/routine_activities.js
@@ -1,8 +1,10 @@
 const express = require("express");
+const { getRoutineById } = require("../db");
 const routine_activitiesRouter = express.Router();
 const {
   updateRoutineActivity,
   destroyRoutineActivity,
+  getRoutineActivityById,
 } = require("../db/routine_activities");
 const { requireUser } = require("./utils");
 
@@ -13,10 +15,9 @@ routine_activitiesRouter.patch(
   async (req, res, next) => {
     const { routineActivityId } = req.params;
 
-    // console.log("ID tester", routineActivityId);
     const { count, duration } = req.body;
 
-    const updateFields = {};
+    const updateFields = { id: routineActivityId };
 
     if (count) {
       updateFields.count = count;
@@ -24,14 +25,16 @@ routine_activitiesRouter.patch(
     if (duration) {
       updateFields.duration = duration;
     }
+
     try {
-      // console.log("Jordan", updateFields);
-      const updatedRA = await updateRoutineActivity(
-        routineActivityId,
-        updateFields
-      );
-      console.log("Pawan test", updatedRA);
-      res.send(updatedRA);
+      const rAQuery = await getRoutineActivityById(routineActivityId);
+      const routineQuery = await getRoutineById(rAQuery.routineId);
+      if (routineQuery.creatorId === req.user.id) {
+        const updatedRA = await updateRoutineActivity(updateFields);
+        res.send(updatedRA);
+      } else {
+        next();
+      }
     } catch (error) {
       next(error);
     }
@@ -43,10 +46,18 @@ routine_activitiesRouter.delete(
   "/:routineActivityId",
   requireUser,
   async (req, res, next) => {
-    const { routineActivityId } = req.params.routineActivityId;
+    const { routineActivityId } = req.params;
     try {
-      const destroyed = await destroyRoutineActivity(routineActivityId);
-      res.send(destroyed);
+      const rAQuery = await getRoutineActivityById(routineActivityId);
+      const routineQuery = await getRoutineById(rAQuery.routineId);
+      if (routineQuery.creatorId === req.user.id) {
+        // console.log("RA TO DELETE ", routineActivityId);
+        const destroyed = await destroyRoutineActivity(routineActivityId);
+        // console.log("RA DELETED ", destroyed);
+        res.send(destroyed);
+      } else {
+        next();
+      }
     } catch (error) {
       next(error);
     }
